refactor: Rest Docs 관련 테스트들은 도메인별 documentation 하위 패키지로 이동
